Modify: add confirm function before an action
when update rootfs and the qemu is running, it is using
the rootfs images, if don't check before this action,
the images is writen by two proc, it may make the images
broken.

So this check is necessary, it will enquire an
input from user, if it is 'y' or 'Y', go to next step.
if not skip the action.